Fix Spark filtering by unprojected columns.
Spark does not request filter columns that are handled by the data
source. However, Spark's Parquet read support cannot handle extra
columns that are not projected, nor reordered columns.

In order to request the filter columns that are not projected from the
Parquet file, or certain schemas with reordered columns, the data source
must return a schema that doesn't match the requested Schema:

* Filter columns must be projected even if not requested by Spark
* Reordered columns must be projected in the file's order

Because the order must match the file, files must all use the same
column order. Currently, this is okay because schema evolution does not
expose a way to reorder columns. This may break on existing data files
that used different column orders.