Spark: Handle identity partition data when reading.
This commit adds support for identity partitions, where data must be
joined from a file's partition tuple to the final row. When identity
partition columns are projected, the Spark reader will split the
requested schema in two, one with columns from the file and one for the
partition data columns. To ensure the final columns are in the correct
order, this adds an unsafe projection to the requested column order.

This also updates handling for filter columns that are not projected.
Before this commit, filter columns that are not required by Spark were
added to the output rows anyway. Now, filter columns are removed by the
final unsafe projection, but are still loaded from the file for filters
to work.