Support int32- and int64-backed decimals for Parquet.
This updates the schema conversion to use int32 and int64 primitive
types for decimal data stored in Parquet. It also adds int, long, and
fixed conversions for decimals in parquet-avro. Avro's built-in decimal
conversions support fixed, but not int32 or int64. To add support, this
replaces Avro's decimal type with a parquet-avro type with the necessary
conversions.