Add bin packing with limited lookback.
This will be used for merging manifest files and for planning splits.

The bin packing algorithm uses two settings, a target size and the
number of bins to consider. When a new bin is added that exceeds the bin
look-back size, the oldest bin is released and removed from
consideration.