Spark: Add attempt number to file names.
This fixes the case where two attempts ask to commit at the same time.
One succeeds and adds the file to the table (if the write succeeds), but
the other aborts and deletes its data file. If both tasks are using the
same file name, the abort deletes the file that gets committed.