Delegate record filtering to Spark.
Instead of using an Iceberg expression filter, return all pushed filters
to Spark as unsupported filters. Spark will apply all of the filters
using codegen.

When Spark supports per-task residuals, it won't be necessary to return
all of the filters. For now, this must assume that any filter could be
in a residual and return them all for Spark to apply.