Fix commit handling when metadata cannot be immediately read.
When refreshing metadata immediately after reading, an update may not be
able to read the metadata due to eventual consistency in S3. This
updates handling in 2 ways:

* Wait up to 10 minutes to load the metadata
* If metadata cannot be loaded but the commit succeeded, continue on

Because the commit succeeded in the second case, no cleanup can be done.
This was already handled in the RuntimeException handler, but accessing
the snapshot for cleanup was throwing a null pointer exception, hitting
the correct cleanup case, but then causing Spark to abort the commit.
When the commit succeeds, no exception should be thrown.