Fix SparkExpressions.convert recursion with nulls.
Conversion returns null to signal that the expression is not supported,
but in recursive calls it wasn't checking that the wrapped expression
was converted to null. This caused some expressions like not(isnan(d))
to be converted and pushed to Iceberg even though they are not
supported. Iceberg later hit a NPE because an expression was null.

This also checks for null expressions in the Expressions factory
methods.