Add support for field documentation. (#59)
This adds an optional documentation string to the field level in Iceberg schemas.

Support for doc strings is added to Spark field conversion and to the UpdateSchema API.

To store field documentation, this adds a "doc" field to each nested field in the JSON representation of types. The spec has also been updated. This is a forward compatible change. Older readers will ignore the new doc field.