Allow schema updates in transactions.
This is safe because schema updates can always read older data. If a
transaction includes a write followed by a schema update, the new schema
can read the data that was just written. Also, writers are allowed to
write with older schemas because the current schema can read any file
written with an older schema. If a transaction includes a schema update
followed by a write, using either the original schema or the new schema
will work.