Fix filtering manifests in unpartitioned tables.
FilteredManifest only ran filters if there was a row filter AND a
partition filter, but it should run filteres if there is a row filter OR
a partition filter.

Because a filter may be null, this also updates the functions that
create evaluators to create an evaluator for alwaysTrue when the
expression is null.