Fix uncommitted file clean-up in transactions.
This adds a callback to delete files and adds a callback in Transaction
that keeps track of deletes instead of running them. When a transaction
is committed, the last set of deletes are run so that the actual deletes
are for the last commit of each operation in the transaction.