Use Spark for Scanning of Manifest Files during Snapshot Expiration
Adds a Spark Action with the aim of parallelizing the manifest scanning
portion of Snapshot expiration.

Previously there was only a single method for expiring old data files associated
with expired Snapshots which required scanning all effected manifests locally. In
order to take advantage of systems which can handle more simultaneous requests and IO
we move the Manifest Scanning portiion of the expiration to Spark. The new functionality
is accesible in a new Spark Action ExpireSnapshotsAction which has a similiar api to
the local task but is instead exeuted on Spark.

The new action is implemented by utilizing the local code to determine which manifest
fils are effected by Snapshot Expiration. Then parallelizing the file names and performing
the scanning of the manifest files remotely. The actual deletion of unneeded data files is still
performed locally.

To get the information required for performing the deletes, the Remove Snapshot class is refactored
so that the methods relating to discovering effected manifests can be called by other modules.