Fix Handling of SparkPartitions with Whitepsace in Location
Previously the listPartition method would use a string representation
of the Partition's URI when creating the Hadoop Path. This constructor would
handle the string as the literal input for the path and ignore any encoded
characters. This casues an issue if the directory or filename has whitespace
or some other special character being reported as it's encoded version
 resulting in FNF exceptions. By switching to passing a URI as
the SparkPartition's location uri, the escaped characters are correctly handled
by the Hadoop Path class..