Address Reviewer Comments
In order to avoid changing the API and SparkSQL compatible types we will fix the whitespace issue by
instead replacing the encoded string representation with a decoded string representation. We use a
method identical to Apache Spark, taking the Hadoop Path representation of the URI and getting the
string representation from that.