Spark: Fix import for paths that include spaces (#1228)
In order to avoid changing the API and SparkSQL compatible types we will fix the whitespace issue by
replacing the encoded string representation with a decoded string representation. We use a
method identical to Apache Spark, taking the Hadoop Path representation of the URI and getting the
string representation from that.