Allow ExpireSnapshots to run without Deleting Files
Previously ExpireSnapshots would always follow the expiration of Snapshots
with a deletion of local DataFiles and Manifests which were no longer relevant
to the Table. This patch introduces an API to skip this deletion phase, which
was always run locally, and instead just expire the snapshots. This allows for
future implementations which can read and remove unused files in parallel or
on distributed frameworks.