Removes Caching from ExpireSnapshotAction
Instead of using a cache to preserve the state from before the
expireSnapshots command, we preserve the table metadata via a
StaticTable reference. This reference doesn't change when the
Snapshosts are expired and allows us to look up all the files
referenced by the prior version of the table without holding
everything in memory.