Fix FileHandle Leak in BaseDataReader for MetadataFiles
The core issue here was that previously when using CloseableIterable.filter
The iterator generated by the wrapped Iterable would be wrapped by a Guava
Iterator regardless of the wrapped iterators closable status. This means
Calling Filter on an iterable would result in a CloseableIterable which
generated incorrectly unclosable Iterators.

To fix this we replace the Guava IterablesFilter operation with the
FilterIterator class. This properly wraps the underlying closable iterator
so that when we generate iterators we still get classes that can close
their wrapped implementations.