WIP - Add an api for Parallelizing Manifest Reading in ManifestGroup
To add distribtued manifest reading to table scans, we allow for a ManifestProcessor
to be used in a TableScan. This class is utilzied when ManifestGroup processes
Manifest files. The default implementation mimics the current code by just wrapping
the reading code in a CloseableIterable. A distributed Spark implementation is also
provided which reads all of the manifests remotely before returning valid entries
for further processing.