Provide API and Implementation for Creating Iceberg Tables from Spark
Previously the SparkUtil class provided an importSparkTable command but
this command suffered from a few shortcomings. It had a difficult api and
did not work directly with DSV2 Iceberg Catalogs. To provide a simpler interface
that shares a similar pattern to currently available Spark Actions we introduce
Actions.migrate and Actions.snapshot.

Through these commands we can both migrate a Table as well as make a snapshot
of an existing table.