Provide Action class for Spark2 and Spark3
Previously there was a sole Action class used by Spark2 and Spark3 modules. This
made it difficult to provide Actions that were only compatible with a single version
of Spark. To fix this we provide a common base class for Actions which are Spark
version agnositic and give each Spark Module an Action class which extends the
common class.

Names, Packages and such remain identical for the end user.