Entries Tables Working with Delete Manifests
Previously the entries metadatat tables would throw an exception if a table
had been modified with a rowDelta and had a delete manifest. To fix this we
switch the reader based on the type of manifest that is being read by the
metadata table so that both type of manifest can be displayed.