Fix Actions to be Compatible with V2 Catalogs
Previously the Action code would use the table name when attempting to look
up MetadataTable names. This is an issue in Spark3 with V2 catalogs because
we strip off the catalog information if the catalog is named "hadoop" or "hive"
and include it in all other cases. When we include the catalog name this breaks
the lookup code which uses the DataSourceRegistry pathway and not the V2 Table
pathway. To fix this we introduce a method for using the V2 pathway when a catalog
name is present.