Fix Avro Pruning Bugs with Empty Structs
Previously we would crash whenever attempting to project only non
data_file columns from a partitioned Iceberg table. This would occur
because our projection in Manifest Readerwould always require the
"data_file" field even if it was an empty struct.

This was an issue because GetProjectedIDs only can retreive ids from
struct fields. If a struct is empty it ignores the field leading to
a projection without the struct.

This worked on unpartitioned tables because of a second bug in the column
pruning which would always included empty structs regardless of if they
were requested or not. This in concert with the above bug would allow un
partitioned tables to work. Their schema data_file schema could never be
empty because it contained a field (partition) which was an empty struct
so it would always be included.

Here we fix both bugs by properly ignoring not requested fields, and
correctly projecting empty structs when they are explicitly requested.