Fix empty struct projection
Inorder to fix all variations of empty struct projections we introduce a second traversal
of the avro schema which just identifies fields which need to be read as empty structs.
We use these integers in parallel with those fields which are being actively included
in the pruned schema, but instead of including them as is, we remove all of their sub-
fields before adding them. Since we are still following in the normal PruneColumns logic
all required parents for the pruned column are added automatically. In addition
new tests are added to make sure that empty and nested empty projections work as expected.