Fix projection of Structs Containing only Unreal Columns
Previously if a user attempted to project a column which did not exist in the file schema
we would throw an error if it was required, or allow it if it was optional. But if this optional
column was the only column in the parent struct, the parent struct would be ignored so
we could not add the child field.

To fix this we add structs which contain only Metadata columns or Optional columsn to our list
of structs to preserve because they may be needed for child fields which don't actually exist.