Fix Hadoop/Hive Catalog Naming Issue
Previously there was a conflict between manually non-Spark-catalogs
and Spark Catalogs in the metadata table name resoultion code. Because
of this ambiguity we will attempt to load all tables as if their catalog
is a named SparkCatalog, if this fails we will fall back to the old
path of removing the "hadoop" or "hive" catalog names and resolving.