When we fail to parse a "CALL" command provide a better error message
Previously we would always delegate to the Spark parser, even when a
CALL request could not be parsed. This leads to confusing errors which
would state that "CALL" was not a good first token. This change has us
only using the custom parser for "CALL" commands and delegating for
everything else.