Spark: Improve extensions parser error messages (#1864)
Previously we would always delegate to the Spark parser, even when a
CALL request could not be parsed. This leads to confusing errors which
would state that "CALL" was not a good first token. This change has us
only using the custom parser for "CALL" commands and delegating for
everything else.