Make lookup of data file by file path robust.
This is a fix for issue #2169.
In `BaseDataReader#getInputFile`, the `InputFile` is looked up in a map
by the path string. The keys in the map are normalized paths. If the
path string in the manifest is not normalized, the lookup will fail
since the strings are not equal. The fix is to normalize the path string
before looking it up in the map.