Adds AddFiles Procedure
This procedure mimics our old support for "IMPORT DATA" but does not allow for dynamic
overwriting of files in partitions. This will now require a seperate DELETE command to
remove a partition. Other than that, the capabilites are identical to the previous functionality
except now the functionality is a procedure rather than an SQL command.