Change Datasource Option Test to not Rely on Hardcoded File Size
Previously this test assumed that a record would always be 611 bytes and the
file created by the test would have roughly twice that many bytes. When parquet
versions are changed this test is broken. Instead we now write a single file
and divide that file size in two to guarentee only 2 partitions will exist if the
test property is correctly applied.