Refactor to use Hive or File methods
Previously we always just used the Spark InMemory file tool to list the partitions in a table
regardless of whether it was a pure file based table or a Hive based table. Now instead we
have two distinct paths, either we have a file based table which is listed like `parquet.path`
or a hive table which has the standard `database.table`. When a table identifier is give for
hive we attempt to read the Catalog's partition listing which lets us discover alternate
partition locations and match the underlying hive catalog.