Switch to Exclusive CommitStateUnknownException
Instead of treating all non CommitFailedExceptions as unknown states we
will explicitly call out when we are unsure of the final commit state. This
means that it is up to the catalogs to accurately and completely enumerate the
failures which indicate that they do not understand the final state of the
commit, but allows for more descriptive exceptions in other cases.