Change execution method for rewrite
Don't use spark.sessionstate.executePlan, this doesn't end up being properly run in the UI.
Instead I used Dataset.ofRows() which properly reports everything to the UI.

Also fixed an issue with the LargeSplitTest. I ran into two issues. One our reader will not split up a row group,
if multiple splits cover the same row group, only one will read (I believe the one that touches the middle most bit)
while all other splits will become no-ops, so I decreased the row group size and increased the number of rows.

The second issue was that the writer only checks to see whether to roll over a file if it is a on a row multiple of 1000,
this means if you ask for a target file size that is smaller than 1000 rows in size you will never split the file.