Reverts Antlr Shadowing and adds Integration Test to Check Proper Shading
Previously we had no test suites which checked if the Spark3Runtime jar was actually
usable with Spark3 builds. To check that the shadowJar is doing the right thing we
add a new integration test which runs using only Spark3 and the ShadowJar. Only a
few tests are included to prove that the jar is stable while we still rely on our
unit tests for the majority of test coverage.