Use `CommitStatusUnknownException` for Nessie
In case the Nessie endpoint did not respond or some other network error that makes it impossible
to detect whether the Nessie server got the request and, more importantly, get the response.

This PR adds a `catch (org.projectnessie.client.http.HttpClientException)` and re-throws it as
the new `CommitStateUnknownException`.

Related to #2328