Example of how we can test Spark3 Artifacts against Both Spark Versions
I removed the previous changes to build.gradle for this demonstration, basically
what I've done is create a second source set which uses the exact same directories
as the normal test configuration. The difference is that the test runtime declares
a dependecy on the newer Spark version while still using classes generated by the
original iceberg-spark3:compile task.

This task will compile the test clases using Spark 3.1 but use the library classes
compiled against Spark 3.0. We can use a similar approach for Spark3-Extensions
and Spark3-Runtime Jar.

One thing we may want to fix here is that I just lazily update the Spark Dependecy
by overriding the value in the base configuration. We may want to exclude and then
reintroduce it, but that seemed a bit more difficult in my testing.