Spark: Fix MergeInto when Source and Target are SinglePartition Distributed
Previously there was an exception when both the source and target of a MERGE INTO
command shared the same distribution, causing an join without an exchange but
partitioning was changed by DynamicFileFilterExec. This could only happen
if both the target and source happend to have a single partition
and were repesented with the distribution SinglePartition. If the filter removes
the partition from the target then there is an unbalance between the number of
partitions but the exchange is also missing. This situation breaks the join.

 We do not currently expose any other distribution information so in most cases
the source and target will return unknown distribution which avoids this problem.
In those situations an exchange always follows the DynamicFileFilterExec which makes
sure the number of partitions always matches for the join.