Refactor state out of Action into Strategy
Previously we used groupID which was exposed state from the RewriteStrategy to
track files as they were rewritten, commits and rewrites would all be referred to
by this string. Now the strategy only returns what files were written and it's up to
the caller to handle all commit behaviors as opposed to previously where the
Spark RewriteHandler was doing the commit work.