Fix Nested Pruning Issue by Exposing Projected Scan Schema
The underlying issue we were trying to solve is that ManifestEntryTable is allowed
to prune columns from the underlying manifest entries that is reads but
it does not expose that it has done so in the Table Schema. Only the ManifestEntries
themselves know they have been pruned and because of this we have no way of recovering
this information at scan time. To fix this we add the ability for DataTasks to expose
a pruned schema which can be used by the various engines to generate proper projections.