Build: Fix deploySite Gradle task
This patch fixes the deploySite Gradle task by ensuring that the
`.asf.yaml` file located in the project root is included in the `site`
directory before it is pushed to the `asf-site` branch. If the
`.asf.yaml` file is not included, ASF INFRA's site publishing mechanism
will ignore the update.

More info is available here:
* https://infra.apache.org/project-site.html
* https://infra.apache.org/asfyaml-mkdocs.html