Address further feedback from Ryan Blue.
Fix a bug in BaseTableScan#useSnapshot.
Some clean up in SnapshotUtil.
Some streamlining in added unit tests.
Refactor spark2 Reader to configure the TableScan on construction,
and let the TableScan get the schema for the snapshot.
Rename new TableIdentifier to SparkTableIdentifier to avoid confusion
with existing TableIdentifier (in different package).
Add convenience constructor to PathIdentifier to avoid modifying tests
for it.