Spark: Compact Medium Size Files (#460)
Previously the rewrite action would ignore files which were below our
target threshold if we could not combine them with a smaller file or piece
of a larger file. This would preclude ever combining the piece of of medium
sized file with several other medium sized files resulting in files of the
correct size. To fix this we attempt to split up all files we are compacting
to give a chance for pieces of medium sized files to be incorporated with each
other.

Previously there was no way of determining the difference between files which
can be split and files which can be split but only at particular offsets. This lead to
some strange planning behavior. To work around that this we expose whether the format
is splittable at offsets or not.