[Benchmark] Improve perf of Dictionary Benchmark Setup
Previously the benchmarking code used a number of "when" constructs to
generate data in rings. This became much more expensive somewhere between
Spark 2.4.5 and 2.4.6 for reasons we are not certain of. In an attempt
to improve performance we have removed all of these constructs and
replaced them with identical modulo operations and add date operations
when applicable.