API: Add function for removing Specs from Metadata.json which are no longer in use
Previously there was no way to remove partition specs from a table once they were
added. To fix this we add an api which searches through all reachable mainfest
files and records their specsIds. Any specIds which do not find are marked for
removal which is done through a serializable commit.