Core: Request Split_Offsets when Planning Scans
Previously we didn't explicitly request the split_offsets column from
manifest files when planning Scan Tasks. This meant that the datafiles
would be split aware when they were divided. To fix this we request the
split_offset information when we do the Manifest read. In addition tests
are added to make sure this isn't missed again in the future.