Core: Throw CommitStateUnknownException in old RewriteDatafilesAction (#2932)
During the commit phase of the old RewriteDatafiles action there is a possibility that the commit fails and the state is unknown. In this case cleaning up newly created files is dangerous and must be avoided. This issue is fixed in the new API but that API is not yet available on all frameworks so the fix is backported here.