Core: Request Split_Offsets when Planning Scans (#3666)
* Core: Request Split_Offsets when Planning Scans

Previously we didn't explicitly request the split_offsets column from
manifest files when planning Scan Tasks. This meant that the datafiles
would not be split aware when they were divided. To fix this we request the
split_offset information when we read manifests. In addition tests
are added to make sure this isn't missed again in the future. Handling of
empty split_offsets is also fixed.