build: apply publishing config after project evaluation
Since we refactored the spark projects to have separate folders/projects
per spark version (in 811af43), gradle was evaluating the maven publish
plugin config (in `deploy.gradle`) before it applied the `shadowJar`
configs required for spark runtime projects, causing the generated POM
to include dependencies when it shouldn't (as we're publishing a shadow
jar).

This fixes it by delaying the maven publishing config to
`afterEvaluate`, when gradle has already applied the shadowJar plugin
for projects.

Fixes #3620